Enhance financial summaries and improve accessibility features across the application
- Added intelligent financial summaries for Manager and Client screens.
- Implemented automatic calculations for current month's receivables and payments.
- Introduced visual alerts for overdue debts in the Client view.
- Modularized code by refactoring components into individual files.
- Migrated financial values from Double to BigDecimal for precise calculations.
- Centralized formatting utilities for monetary values and dates.
- Improved accessibility by adding content descriptions and enhancing touch targets.

+++ app/src/main/java/br/com/lucolimac/xuxubank/domain/model/Debt.kt
@@ -1,12 +1,13 @@
 package br.com.lucolimac.xuxubank.domain.model
 
 import br.com.lucolimac.xuxubank.data.local.entity.DebtStatus
+import java.math.BigDecimal
 
 data class Debt(
     val id: Long,
     val clientId: Long,
     val description: String,
-    val amount: Double,
+    val amount: BigDecimal,
     val dueDate: Long?,
     val status: DebtStatus,
     val totalInstallments: Int,
